rugby-ai-predictor/requirements.txt
firebase-functions==0.1.0
firebase-admin==6.2.0
google-cloud-firestore>=2.14.0
google-cloud-storage>=2.14.0
google-cloud-secret-manager>=2.16.0
protobuf>=5.0.0
numpy>=1.24.0,<2.0.0
pandas>=2.0.0,<2.3.0
scikit-learn==1.3.2
joblib>=1.3.0
xgboost>=2.0.0
torch
requests>=2.32.0
python-dotenv>=1.0.0
pytz>=2024.1
tenacity>=8.2.0,<9.0.0
pydantic>=2.9.0
